requirements.txt
Flask
pandas
numpy
faiss-cpu
requests
gunicorn